【版本号】V1.0.6.11 【变更类型】新增特性 【代码描述】1.修改确认订单bean的idPerson类型为long

+++ Mall/app/src/main/java/com/giveu/shoppingmall/model/bean/response/ConfirmOrderScResponse.java
@@ -15,7 +15,7 @@ public class ConfirmOrderScResponse extends BaseBean<ConfirmOrderScResponse> {
      * payMoney : 4399.0
      */
 
-    public int idPerson;
+    public long idPerson;
     public String orderNo;
     public String payMoney;
     public String payType;
